refactor: Rm genetic & upd stats (#58)
* refactor: Rm a lot of genetic code

* refactor: Rm genetic config&constants

* refactor: Replace stats with one func

* refactor: Replace old genetic stats with new

* refactor: Finalize old stats removal

* chore: Add 'VSharp: ' to maps, upd code

---------
