+++ AIAgent/connection/broker_conn/requests.py
@@ -3,11 +3,10 @@ import logging
 from urllib.parse import urlencode
 
 import httplib2
-
 from common.classes import SVMInfo
-from common.constants import ResultsHandlerLinks, WebsocketSourceLinks
+from common.constants import WebsocketSourceLinks
 
-from .classes import Agent2ResultsOnMaps, ServerInstanceInfo
+from .classes import ServerInstanceInfo
 
 
 def acquire_instance(svm_info: SVMInfo) -> ServerInstanceInfo:
@@ -38,24 +37,3 @@ def return_instance(instance: ServerInstanceInfo):
     else:
         logging.error(f"{response.status} on returning {instance}")
         raise RuntimeError(f"Not ok response: {response.status}")
-
-
-def send_game_results(data: Agent2ResultsOnMaps):
-    response, content = httplib2.Http().request(
-        ResultsHandlerLinks.POST_RES,
-        method="POST",
-        body=data.to_json(),
-    )
-
-    if response.status == 200:
-        logging.info(f"map2result was sent")
-    else:
-        logging.error(f"{response.status} on sending map2result: {data}")
-        raise RuntimeError(f"Not ok response: {response.status}")
-
-
-def recv_game_result_list() -> str:
-    response, content = httplib2.Http().request(ResultsHandlerLinks.GET_RES)
-    games_data = content.decode("utf-8")
-    logging.info(f"Acquired games data")
-    return games_data
