+++ AIAgent/epochs_statistics.py
@@ -1,12 +1,16 @@
-from dataclasses import dataclass
 import multiprocessing as mp
+from dataclasses import dataclass
 from multiprocessing.managers import BaseManager
 from pathlib import Path
-import pandas as pd
 from statistics import mean
+from typing import TypeAlias
+
 import natsort
-from epochs_statistics.tables import table_to_string
-from epochs_statistics.common import EpochNumber, SVMName
+import pandas as pd
+from common.classes import Map2Result
+
+EpochNumber: TypeAlias = int
+SVMName: TypeAlias = str
 
 
 def sort_dict(d):
@@ -64,10 +68,12 @@ class StatisticsCollector:
         epoch: EpochNumber,
         SVM_name: SVMName,
         average_result: float,
-        df: pd.DataFrame,
+        map2results_list: list[Map2Result],
     ):
         results = self._sessions_info.get(epoch, {})
-        results[SVM_name] = StatsWithTable(average_result, df)
+        results[SVM_name] = StatsWithTable(
+            average_result, convert_to_df(SVM_name, map2results_list)
+        )
         self._sessions_info[epoch] = sort_dict(results)
         self._update_file()
 
@@ -103,7 +109,7 @@ class StatisticsCollector:
                     )
                 )
             )
-            epochs_results += table_to_string(df) + "\n"
+            epochs_results += df.to_markdown(tablefmt="psql") + "\n"
         return epochs_results
 
     def _update_file(self):
@@ -113,6 +119,20 @@ class StatisticsCollector:
         with open(self._file, "w") as f:
             f.write(SVMS_info)
             f.write(epochs_results)
+
+
+def convert_to_df(svm_name: SVMName, map2result_list: list[Map2Result]) -> pd.DataFrame:
+    maps = []
+    results = []
+    for map2result in map2result_list:
+        map_name = map2result.map.MapName
+        game_result_str = map2result.game_result.printable(verbose=True)
+        maps.append(f"{svm_name} : {map_name}")
+        results.append(game_result_str)
+
+    df = pd.DataFrame(results, columns=["Game result"], index=maps).T
+
+    return df
 
 
 class StatisticsManager(BaseManager):
