--- AIAgent/learning/selection/mutation_type.py
@@ -1,12 +1,0 @@
-from enum import Enum
-
-
-class MutationType(Enum):
-    RANDOM = "random"
-    SWAP = "swap"
-    INVERSION = "inversion"
-    SCRAMBLE = "scramble"
-    ADAPTIVE = "adaptive"
-
-    def __get__(self, instance, owner):
-        return self.value
